requirements.txt
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2
mysqlclient==1.4.1
django-environ==0.8.1
google-api-python-client==2.55.0
djangorestframework==3.13.1
django-filter==21.1
django-taggit==3.0.0
django-crontab==0.7.1